requirements.txt
# Основные компоненты API
fastapi==0.115.14
uvicorn==0.35.0
requests==2.32.4
httpx==0.28.1

# LangChain - ядро для агента
langchain==0.2.10
langgraph==0.1.1
langchain-core==0.2.23
langchain-community==0.2.10
langchain_openai==0.1.17
langchain-huggingface==0.0.1

# Векторная база данных и эмбеддинги
lancedb==0.24.2
sentence-transformers==3.0.1

# Конфигурация и валидация
PyYAML==6.0.2
pydantic==2.11.7
pydantic-settings==2.3.4

# Интерфейсы
python-telegram-bot==22.2
numpy<2.3

# Для тестов
pytest==8.4.1